refactor: refactor GatewayStreamer place
